requirements.txt
attrs==21.4.0
Automat==20.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
filelock==3.6.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.5.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.8.0
oauthlib==3.2.0
parsel==1.6.0
Protego==0.2.1
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.7
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
tldextract==3.2.0
Twisted==22.2.0
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
w3lib==1.22.0
zope.interface==5.4.0
